requirement.txt
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
InstructorEmbedding
sentence_transformers